requirements.txt
faster-whisper 
ffmpeg-python 
pytube 
moviepy